+++ test/fhirpath.test.js
@@ -106,10 +106,10 @@ const generateTest = (test, testResource) => {
     if (!test.error) {
       expect(exception).isNotError();
       // Run the result through JSON so the FP_Type quantities get converted to
-      // strings.  Also , if the result is an FP_DateTime, convert to a Date
-      // object so that timezone differences are handled.
-      if (result.length === 1 && result[0] instanceof FP_DateTime)
-        expect(new Date(result[0])).toEqual(new Date(test.result[0]))
+      // strings.  Also, if the result is an FP_DateTime, use compare() so that
+      // timezone differences and leap seconds are handled.
+      if (result.length === 1 && test.result.length === 1 && result[0] instanceof FP_DateTime)
+        expect(FP_DateTime.checkString(result[0].ctx, test.result[0]).compare(result[0])).toEqual(0)
       else
         expect(JSON.parse(util.toJSON(result))).toEqual(test.result);
     }